asdf

+++ src/Arrangementvindu.java
@@ -19,7 +19,7 @@ public class Arrangementvindu extends JApplet {
 	private static final long serialVersionUID = 1L;
 	private JLabel placeholder;
 	private JTextField navnFelt, beskFelt, prisFelt, refFelt,altFelt1,altFelt2,bildeNavnFelt;
-	private JButton finnKnapp, slettKnapp, regKnapp, listeKnapp,bildeKnapp;
+	private JButton finnKnapp, slettKnapp, regKnapp, listeKnapp,bildeKnapp,oppdaterKnapp;
 	private JTextArea tekstområde;
 	private JScrollPane utskriftområde;
 	private BorderLayout layout,centerLayout,centerPageStartLayout;
@@ -41,6 +41,7 @@ public class Arrangementvindu extends JApplet {
 	private Kalenderpanel kalenderpanel;
 	private Date dato;
 	public Personregister preg;
+	private File bildeFil;
 	
 	private String[] ekstraInputL() { // Sørger for å legge til nye unike oppføringer av lokaler.
 		HashSet<String> a = new HashSet<>(Arrays.asList(k.lokalListe()));
@@ -64,15 +65,13 @@ public class Arrangementvindu extends JApplet {
 		Lokale lok = k.finnType(l);
 		
 		if (l.equalsIgnoreCase("oppdater lokal-liste")) {
-			north.setLayout(new GridLayout(7, 1)); // 7 rows 2 columns; no gaps);
 			
 			for (String s : ekstraInputL()) {
 				lokalvelger.addItem(s);
 			}
 			lokalvelger.revalidate();
 			System.out.println(ekstraInputL());
-			north.add(new JLabel(""));
-			north.add(new JLabel("Lokal-liste er oppdatert!"));
+			tekstområde.setText("Liste over arrangementer er oppdatert!");
 		}
 		else if (lok instanceof Cafe) {
 			north.setLayout(new GridLayout(7, 2)); // 5 rows 2 columns; no gaps);
@@ -118,14 +117,11 @@ public class Arrangementvindu extends JApplet {
 	private void addSpecificK(String k) {
 		
 		if (k.equalsIgnoreCase("oppdater kontaktliste")) {
-			north.setLayout(new GridLayout(7, 1)); // 7 rows 2 columns; no gaps);
-			
 			for (String s : ekstraInputK()) {
 				kontaktvelger.addItem(s);
 			}
 			kontaktvelger.revalidate();
-			north.add(new JLabel(""));
-			north.add(new JLabel("Liste over kontaktpersoner er oppdatert!"));
+			tekstområde.setText("Liste over kontaktpersoner er oppdatert!");
 		}
 		
 	}
@@ -152,7 +148,11 @@ public class Arrangementvindu extends JApplet {
 	}
 	public boolean slettFil(Arrangement a) {
     	try{
-    		 
+    		
+    		String bildeSti = a.get_bildeSti();
+    		if (bildeSti==null) {
+    			return false;
+    		}
     		File file = new File(a.get_bildeSti());
  
     		if(file.delete()){
@@ -220,6 +220,8 @@ public class Arrangementvindu extends JApplet {
 			regKnapp = new JButton( "Registrer arrangement" );
 			listeKnapp = new JButton( "List arrangement" );
 			bildeKnapp = new JButton("Last inn bilde");
+			bildeNavnFelt.setText("");
+			oppdaterKnapp = new JButton("Oppdater Arrangement");
 			
 			//////////////////////////////////////////
 			/////////// GUI LAYOUT START /////////////
@@ -229,7 +231,7 @@ public class Arrangementvindu extends JApplet {
 			centerLayout = new BorderLayout(5,5);
 			centerPageStartLayout = new BorderLayout(6,5);
 			
-			bottomGrid = new GridLayout(1, 4);
+			bottomGrid = new GridLayout(1, 5);
 			topGrid = new GridLayout(5, 2);
 			centerBot = new GridLayout(1,1);
 			// DECLARATIONS END
@@ -291,6 +293,7 @@ public class Arrangementvindu extends JApplet {
 			south.add(slettKnapp);
 			south.add(regKnapp);
 			south.add(listeKnapp);
+			south.add(oppdaterKnapp);
 			// BOTTOM GRID END
 			
 			c = getContentPane();
@@ -314,7 +317,8 @@ public class Arrangementvindu extends JApplet {
 			lokalvelger.addActionListener( lytter );
 			kontaktvelger.addActionListener( lytter );
 			listeKnapp.addActionListener(lytter);
-			bildeKnapp.addActionListener(lytter);			
+			bildeKnapp.addActionListener(lytter);
+			oppdaterKnapp.addActionListener(lytter);
 			
 			checkbox.addItemListener(tLytter);
 			
@@ -466,7 +470,7 @@ public class Arrangementvindu extends JApplet {
 	      else if ( e.getSource() == bildeKnapp ) {
 	    	  try {
 	      		bildehandler = new Bildehandler();
-	      		File bildeFil = bildehandler.hentFil();
+	      		bildeFil = bildehandler.hentFil();
 	      		bilde = ImageIO.read(bildeFil);
 	      		bildeIcon.setImage(bilde);
 	      		bildeNavnFelt.setText(bildeFil.getName());
@@ -475,6 +479,60 @@ public class Arrangementvindu extends JApplet {
 	    	  } catch(Exception ex) {
 	    		  tekstområde.setText("Noe gikk galt.");
 	    	  }
+	      }
+	      else if ( e.getSource() == oppdaterKnapp ) {
+	    	  String navn = navnFelt.getText();
+	    	  String besk = beskFelt.getText();
+	    	  String alt = altFelt1.getText();
+	    	  String alt2 = altFelt2.getText();
+	    	  String pris = prisFelt.getText();
+	    	  String bildeB = bildeNavnFelt.getText();
+	    	  int refNr = Integer.parseInt(refFelt.getText());
+	    	  Lokale lokFunnet = k.arrangementViaK(refNr);
+	    	  Arrangement arrFunnet = lokFunnet.finnArrangement(refNr);
+	    	  if (!navn.equals(""))
+	    		  arrFunnet.set_Navn(navn);
+	    	  if (!besk.equals(""))
+	    		  arrFunnet.set_Beskrivelse(besk);
+	    	  if (!alt.equals(""))
+	    		  arrFunnet.set_Info(alt);
+	    	  if (!alt2.equals(""))
+	    		  arrFunnet.set_Info2(alt2);
+	    	  if (!pris.equals(""))
+	    		  arrFunnet.set_Pris(Integer.parseInt(pris));
+	    	  if (!bildeB.equals("")) {
+					if (arrFunnet != null) {
+						Object[] options = {"Ja",
+						                    "Nei",};
+						int n = JOptionPane.showOptionDialog(null,
+						    "Vil du erstatte bildet med et nytt?",
+						    "Hei bruker.",
+						    JOptionPane.OK_CANCEL_OPTION,
+						    JOptionPane.INFORMATION_MESSAGE,
+						    null,
+						    options,
+						    options[0]);
+						if (n == 1) {
+							tekstområde.setText("* Bildet ble ikke erstattet *\r\n");
+							tekstområde.append(arrFunnet.toString());
+			    	    	clearFields();
+							return;
+						}
+						slettFil(arrFunnet);
+						String bildenavn = "./images/"+arrFunnet.get_Navn()+"-"+"bilde.png";
+	    				File outputfile = new File(bildenavn);
+	    	    		arrFunnet.set_Bildesti(bildenavn);
+	    				try {
+							ImageIO.write(bilde, "png", outputfile);
+						} catch (IOException e1) {
+							e1.printStackTrace();
+						};
+					}
+	    	  }
+
+	    	  tekstområde.setText(arrFunnet.toString());
+	    	  clearFields();
+	    	  setPlaceHolderImg();
 	      }
 	      
 		    int n = lokalvelger.getSelectedIndex();
